% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_add.pointyear.R
\name{add.pointyear}
\alias{add.pointyear}
\title{Adds year and pointyear columns to data}
\usage{
add.pointyear(data)
}
\arguments{
\item{data}{A dataframe object}
}
\value{
Dataframe with added columns
}
\description{
For veg data, formats column with date as a date object in R, adds a column for year as a numeric object, and pointyear as a character object that inludes the point name and the survey year

Data supplied needs to have column named "PointId" or "Point.Id" and a date column named "Date" or "Event.Date"

Makes use of package function read.date()
}
\examples{
data = add.pointyear(data)

}
