% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.zeros.r
\name{add.zeros}
\alias{add.zeros}
\title{Adds zeros to dataframe for statistical analysis}
\usage{
add.zeros(sum.data)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A data frame that should be saved as newpc2
}
\description{
Takes bird data that has been prepared by bird_prepare function and turned into species by bird_species.all and merged with visits dataframe created by bird_visits.
   The data entered into this function should be created by: pc<-merge(species, visits, by="PointYear", all=TRUE) pc<-droplevels(pc)
}
\examples{
newpc2 = bird_species.common(pc)

}
