% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_remove_first.r
\name{all_remove.first}
\alias{all_remove.first}
\title{Remove first row of data.frame if empty}
\usage{
all_remove.first(df)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
A ggplot of functional cover with error bars
}
\description{
A reoccuring problem when importing data frames into rstudio is a blank first row that is added to the data frame. This function looks at the firsts row to see if the first cell has an empty value. If it does then it will removes the row and return the dataset.
}
\examples{
data = functional.cover.plot(lpi, type = "absolute")

}
