% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_Tidyname.R
\name{all_Tidynames}
\alias{all_Tidynames}
\title{Tidy Column Names of data frame}
\usage{
all_Tidynames(df)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A data frame with tidy names
}
\description{
For any data frame, changes column names if they are not unique or tidy.
}
\examples{
data = all_Tidynames(df)

}
