% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_birdcount.year.r
\name{bird_birdcount.year}
\alias{bird_birdcount.year}
\title{Create Species List}
\usage{
bird_birdcount.year(df, transect,
  surveyyear = c(levels(as.factor(df$YEAR))))
}
\arguments{
\item{df}{A data frame object. Only will take species which is created by:  bird_species.common(). See that function for previous steps.}

\item{transect}{A ranch code or a list of ranch codes ie."TOKA".}

\item{surveyyear}{A year or multiple years. ie. c(2016,2018)}
}
\value{
A data frame that summerises counts of each bird species per year.
}
\description{
Takes bird data that has been prepared by bird_prepare function and bird_species.common function to create a bird list that can be controlled by transect and surveyyear.
}
\examples{
data = bird_birdcount.year(species, "TOKA", 2016)

}
