% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_convert.distance.r
\name{bird_convert.distance}
\alias{bird_convert.distance}
\title{Convert distance bins into continuous values}
\usage{
bird_convert.distance(df)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
Point Count data with distance.bin column as numeric.
}
\description{
The Rangeland Monitoring Program has a few different methods for collecting bird point count data: VCP300, VCP10_30, VCP25, and FR50. To analyze these datasets together the distance bin values (aka character strings) need to be converted to numbers (integers). This is useful for data collected using VCP10_30, VCP25, or FR50 protocols.
This was done using the mapvalues function and assigning a number value to all bins. I averaged the max and min values for each distance bin to determine the number value. This should only be used to analyzing point count data as presence/absence and should not be used for density analysis.
}
\examples{
data = bird_convert.distance(data)

}
