% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_focal.graph.r
\name{bird_focal.graph}
\alias{bird_focal.graph}
\title{Create a Focal Species Abundance per Year Graph}
\usage{
bird_focal.graph(df, transect = c(levels(as.factor(df$Transect))),
  surveyyear = c(levels(as.factor(df$YEAR))),
  choose_focal_group = c("Grassland", "Oak.Woodland", "Riparian"))
}
\arguments{
\item{df}{A data frame object. Only will take wide.data which is created by: add.zero function , bird_wide.data(), bird_prepare(). See those functions for additional previous steps.}

\item{transect}{A ranch code or a list of ranch codes ie."TOKA".}

\item{surveyyear}{A year or multiple years. ie. c(2016,2018)}

\item{choose_focal_group}{Three bird focal groups: Grassland, Oak Woodland, and Riparian. Each contain Spp codes that we consider focal species for that habitat type.}
}
\value{
A graph that summerises focal species abundance per year.
}
\description{
Takes bird data that has been prepared by bird_prepare function, add.zero function and bird_wide.data function to create a focal species per year graph based on ranch transect and focal group
}
\examples{
bird_focal.graph(wide.data, transect="TOKA", surveyyear= 2016, choose_focal_group="Grassland")

}
