% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_MAPdistance.bin.r
\name{bird_MAPdistance.bin}
\alias{bird_MAPdistance.bin}
\title{Replaces Distance Bin Values from different RMN point count protocols.}
\usage{
bird_MAPdistance.bin(df)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
A dataframe with the column Distance.bin as numeric values
}
\description{
Replaces distance bin Values from character strings to number values. There are four different RMN point count protocols with some resulting in numeric numbers and some wiht character strings (10-20). When analyzing data from multiple protocols you need to account for these differences.
}
\examples{
data = bird_MAPdistance.bin(df)

}
