% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_prep.columns.r
\name{bird_prep.columns}
\alias{bird_prep.columns}
\title{Adds columns to bird dataframe}
\usage{
bird_prep.columns(df)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
A dataframe prepared for data analysis with added columns.
}
\description{
Adds columns YEAR, SURVEY, Pointyear,SAMPLE and Tally
}
\examples{
data = bird_prep.columns(df)

}
