% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_prepare.r
\name{bird_prepare}
\alias{bird_prepare}
\title{Prepares bird data for analysis}
\usage{
bird_prepare(df, x)
}
\arguments{
\item{df}{is a dataframe. Will take raw Point Count CSV files from Point Blue's CADC website}

\item{x}{is the distance you want to subset the distance bin. Same as distance in bird_subset.distance(df, distance).}
}
\value{
Cleaned and subsetted dataframe. For consistancy save as df2
}
\description{
First step in the data cleaning process. Consists of 6 functions that clean and subset the data.
}
\examples{
df2<-bird_prepare(df)

}
