% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_species.common.r
\name{bird_species.common}
\alias{bird_species.common}
\title{Creates Species dataframe with common names}
\usage{
bird_species.common(df)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A data frame that should be saved as species2
}
\description{
Takes bird data that has been prepared by bird_prepare function and creates a dataframe summerising all observations of species by year. Different than bird_species.all because it adds common and scientific names of birds.
}
\examples{
species2 = bird_species.common(df)

}
