% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_species.list.r
\name{bird_species.list}
\alias{bird_species.list}
\title{Create Species List with common names}
\usage{
bird_species.list(df, transect = c(levels(as.factor(df$Transect))),
  surveyyear = c(levels(as.factor(df$YEAR))))
}
\arguments{
\item{df}{A data frame object. Only will take species2 which is created by:  bird_species.common(). See that function for previous steps.}

\item{transect}{A ranch code or a list of ranch codes ie."TOKA".}

\item{surveyyear}{A year or multiple years. ie. c(2016,2018)}
}
\value{
A data frame that summerises counts of each bird species per year.
}
\description{
Takes bird data that has been prepared by bird_prepare function and bird_species.common function to create a bird list that can be controlled by transect and surveyyear.
}
\examples{
data = bird_species.list(species2, "TOKA", 2016)

}
