% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_subset.cue.r
\name{bird_subset.cue}
\alias{bird_subset.cue}
\title{Subset bird data based on detection cue}
\usage{
bird_subset.cue(df)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A data frame that is subsetted based on detection
}
\description{
Takes bird data and removes observation of flyovers, juveniles, and certain distance bins.
}
\examples{
data = bird_subset.cue(df)

}
