% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_visits.r
\name{bird_visits}
\alias{bird_visits}
\title{Create Visit Table}
\usage{
bird_visits(df)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A data frame with visits
}
\description{
Takes bird data that has been prepared by bird_prepare function and creates a dataframe of visits per point.
}
\examples{
data = bird_visits(df)

}
