% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_wide.data.r
\name{bird_wide.data}
\alias{bird_wide.data}
\title{Creates wide data frame for Point Count data}
\usage{
bird_wide.data(df)
}
\arguments{
\item{df}{is a dataframe. Only works with newpc2 created by add.zeros() and it's previous steps.}
}
\value{
Data frame with the is the wide version of newpc2 with added column of Richness
}
\description{
For the output of add.zeros(). This function turns the dataframe from a long formate to wide formate and adds a richness column.
}
\examples{
bird_wide.data(newpc2)

}
