% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_wide.focal.r
\name{bird_wide.focal}
\alias{bird_wide.focal}
\title{Creates a subsetted wide data frame for Point Count data}
\usage{
bird_wide.focal(df, transect = c(levels(as.factor(df$Transect))),
  surveyyear = c(levels(as.factor(df$YEAR))),
  choose_focal_group = c("Grassland", "Oak Woodland", "Riparian"))
}
\arguments{
\item{df}{is a dataframe. Only works with newpc2 created by add.zeros() and it's previous steps.}

\item{transect}{A ranch code or a list of ranch codes ie."TOKA".}

\item{surveyyear}{A year or multiple years. ie. c(2016,2018)}

\item{choose_focal_group}{Three bird focal groups: Grassland, Oak Woodland, and Riparian. Each contain Spp codes that we consider focal species for that habitat type.}
}
\value{
A subsetted wide data frame by focal group
}
\description{
For the output of bird_wide.data(). This function subsets the wide dataframe by ranch, year and focal species.
}
\examples{
bird_wide.focal(newpc2, "TOKA", surveyyear= 2018, choose_focal_group="Grassland")

}
