% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_compaction.plot.r
\name{compaction.plot}
\alias{compaction.plot}
\title{Make a soil compaction plot}
\usage{
compaction.plot(data, transect, labels = TRUE, guide = FALSE)
}
\arguments{
\item{guide}{}
}
\value{
Shows soil compaction plot
}
\description{
From output of prepare.soil.triangle() creates a soil compaction plot displaying BD and Infiltration data, and compares to other ranches in the region
}
\examples{
compaction.plot(soil, "ranchname")


}
