% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_cover.sum.plot.R
\name{cover.sum.plot}
\alias{cover.sum.plot}
\title{Plot species richness and cover}
\usage{
cover.sum.plot(lpi, releve, transect,
  surveyyear = c(levels(as.factor(lpi$year)),
  levels(as.factor(checklist$year))),
  choose.variable = c("SpeciesRichness", "Litter", "Thatch",
  "BareGround", "Trees", "Shrubs"), plot.names = c("Species Richness",
  "Litter", "Thatch", "Bare Ground", "Trees", "Shrubs"),
  multiple_years = TRUE, barcolors = c("olivedrab", "steelblue3",
  "mediumpurple4", "orange3", "tan2"), legend = TRUE,
  legendname = "Year", xlab = "Point ID", ylab = "Value",
  xangle = 45)
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{releve}{A dataframe object of releve data from a veg survey}

\item{transect}{Ranch for which to make the plot}

\item{surveyyear}{The years for which to make the table}

\item{choose.variable}{Character vector that identifies which variables to summarize. See cover.summary()}

\item{multiple_years}{Boolean that determines whether multiple years of data are plotted. if FALSE, only most recent year is plotted}

\item{barcolors}{Character vector of bar fill colors for each year of data}

\item{legend}{Boolean that specifies whether to display a legend}

\item{legendname}{A title for your legend}

\item{xlab, ylab}{Labels for axes}

\item{xangle}{Angle to display point names on x axis}

\item{plot.name}{Character vector of names of plots. If choose.variable is manually changed, then plot.names also needs to be changed.}
}
\value{
ggplot object
}
\description{
For veg data, plots cover of shrubs, trees, litter, thatch, bare ground. Tree and shrub covers come from releve estimates
}
\examples{
data = cover.sum.plot(lpi, checklist)

}
