% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_cover.summary.R
\name{cover.summary}
\alias{cover.summary}
\title{Adds year and pointyear columns to data}
\usage{
cover.summary(checklist, lpi)
}
\arguments{
\item{checklist}{A dataframe object of checklist data from a veg survey}

\item{lpi}{A dataframe object of lpi data from a veg survey}
}
\value{
A summary of cover
}
\description{
For veg data, formats column with date as a date object in R, adds a column for year as a numeric object, and pointyear as a character object that inludes the point name and the survey year

Data supplied needs to have column named "PointId" or "Point.Id" and a date column named "Date" or "Event.Date"

Makes use of package function read.date()
}
\examples{
data = cover.summary(checklist, lpi)

}
