% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_data.sum.R
\name{data.sum}
\alias{data.sum}
\title{Make a table to summarize data}
\usage{
data.sum(data, transect, surveyyear = levels(as.factor(data$YEAR)),
  return_df = FALSE)
}
\arguments{
\item{data}{dataframe, output of soil.final.cleanup()}

\item{transect}{Character string of selected ranch code}

\item{surveyyear}{Years for which to use data}
}
\value{
Shows soil data in a nice-looking table

data summary table
}
\description{
From output of soil.final.cleanup() creates a soil texture plot and compares a ranch to other ranches in the dataset
}
\examples{
data.sum(soil, "ranchname")


}
