% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_inf.R
\name{extrapolate_inf}
\alias{extrapolate_inf}
\title{Extrapolate Infiltration Time}
\usage{
extrapolate_inf(height, time = 45)
}
\arguments{
\item{time}{}
}
\value{
Extrapolated time of Infiltration
}
\description{
For infiltration tests that were stopped before completely finishing, this funtion uses the remaining depth (cm) and the run time (minutes) of the test to extrapolate the time of a complete infiltration.
}
\examples{
extrapolate_inf(1.5, 50)

}
