% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_functional.cover.change.plot.R
\name{functional.cover.change.plot}
\alias{functional.cover.change.plot}
\title{Functional cover change summary plot}
\usage{
functional.cover.change.plot(lpi,
  surveyyear = levels(as.factor(lpi$year)), type = "absolute",
  xlab = "Functional Group", ylab = paste("Cover Change", " ",
  min(levels(abs$year)), "-", max(levels(abs$year)), sep = ""))
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{surveyyear}{The years for which to make the plot}

\item{type}{"absolute" or "relative" cover}

\item{xlab, ylab}{Axis Labels}
}
\value{
A ggplot of functional cover change with error bars
}
\description{
For veg data, summarizes change in mean cover of functional groups for a property
}
\examples{
data = functional.cover.plot(lpi, type = "absolute")

}
