% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_functional.cover.change.table.R
\name{functional.cover.change.table}
\alias{functional.cover.change.table}
\title{Functional cover change summary table}
\usage{
functional.cover.change.table(lpi, transect, invasives = FALSE,
  surveyyear = levels(as.factor(lpi$year)), type = "absolute",
  casted = TRUE)
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{transect}{Ranch for which to make plot}

\item{invasives}{Boolean whether to include invasives as a functional group}

\item{surveyyear}{The years for which to make the plot}

\item{type}{"absolute" or "relative" cover}

\item{casted}{Boolean whether to return a cast dataframe. FALSE returns a molten data frame}

\item{xlab, ylab}{Axis Labels}
}
\value{
A ggplot of functional cover change with error bars
}
\description{
For veg data, summarizes change in mean cover of functional groups for each point on a property
}
\examples{
data = functional.cover.plot(lpi, type = "absolute")

}
