% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_functional.cover.table.R
\name{functional.cover.table}
\alias{functional.cover.table}
\title{Create a table of cover for functional groups at each point}
\usage{
functional.cover.table(lpi, transect, type = "absolute",
  invasives = FALSE, surveyyear = max(levels(as.factor(lpi$year))),
  includemeta = FALSE)
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{transect}{Ranch for which to make cover table}

\item{type}{Type of cover to calculate - either "absolute" or "relative"}

\item{invasives}{Boolean that specifies whether to count invasive species as a separate functional group}

\item{surveyyear}{Year for which to calculate cover}

\item{includemeta}{=FALSE Boolean on whether to include columns specifying Transect and pointyear. Useful if using for further analysis but unecessary for display}
}
\value{
A dataframe summary of cover of functional groups
}
\description{
For veg data, summarizes cover of each functional group at each point.
}
\examples{
data = functional.cover.table(lpi, type = "absolute")

}
