% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_long.lpi.R
\name{long.lpi}
\alias{long.lpi}
\title{Summarize number of lpi hits by species}
\usage{
long.lpi(lpi)
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}
}
\value{
A dataframe summary of species hits at each survey
}
\description{
For veg data, summarizes number of hits of each species at each point. Intended as intermediate data prep step
}
\examples{
summart = lpi.summary(lpi)


}
