% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_lpi.summary.R
\name{lpi.summary}
\alias{lpi.summary}
\title{Summarize number of lpi hits by species}
\usage{
lpi.summary(lpi)
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}
}
\value{
A dataframe summary of species hits at each survey
}
\description{
For veg data, summarizes number of hits of each species at each point. Intended as intermediate data prep step
}
\examples{
summart = lpi.summary(lpi)

}
