% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_pair.data.R
\name{pair.data}
\alias{pair.data}
\title{Aggregate Soil Data by Point}
\usage{
pair.data(newdata, singlesurvey = TRUE)
}
\arguments{
\item{singlesurvey}{}
}
\value{
Dataframe with soil data from different years merged by point. Columns are added to display percent change in soil data between surveys
}
\description{
Creates a dataframe where each soil point has one row. Each data column is duplicated for each survey event.

First argument has to be the output of the function survey.sums()
}
\examples{
pair.data(data, singlesurvey = TRUE)
paired_data = pair.data(soildata, singlesurvey = FALSE)

}
