% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_percent.change.plot.R
\name{percent.change.plot}
\alias{percent.change.plot}
\title{Make a percent change boxplot of soil carbon and bulk Density}
\usage{
percent.change.plot(data, transect)
}
\arguments{
\item{guide}{}
}
\value{
Shows soil compaction plot
}
\description{
Creates a plot that shows average change in SOC at 2 depths and BD compared to regional average
}
\examples{
compaction.plot(soil, "ranchname")


}
