% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_points.cover.plot.R
\name{points.cover.plot}
\alias{points.cover.plot}
\title{Create a plot of cover for functional groups at each point}
\usage{
\method{points}{cover.plot}(lpi, type = "absolute", invasives = FALSE,
  legend.position = "top", pallete = "YlGnBu", x.angle = 45)
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{type}{Type of cover to calculate - either "absolute" or "relative"}

\item{invasives}{Boolean that specifies whether to count invasive species as a separate functional group}

\item{legend.position}{string specifying legend position}

\item{pallete}{String identifying color scheme passed to scale_fill_brewer}

\item{x.angle}{Angle of x axis table}
}
\value{
A plot of functional group cover at each point in a dataset
}
\description{
For veg data, summarizes cover of each functional group at each point.
}
\examples{
data = points.cover.plot(lpi, type = "absolute")

}
