% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_read.date.R
\name{read.date}
\alias{read.date}
\title{Interpret data}
\usage{
read.date(date_string)
}
\arguments{
\item{date_string}{}
}
\value{
date object
}
\description{
Interprets correctly formatted string as a date
}
