% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_remove_first.r
\name{remove_first}
\alias{remove_first}
\title{Remove first row if empty}
\usage{
remove_first(df)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
A ggplot of functional cover with error bars
}
\description{
For veg data, summarizes and plots mean cover of functional groups for a property
}
\examples{
data = functional.cover.plot(lpi, type = "absolute")

}
