% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate standard error}
\usage{
se(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Standard error (numeric)
}
\description{
Calculate standard error of a numeric vector
}
\examples{
se(x)

}
