% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.final.cleanup.R,
%   R/soil_soil.final.cleanup&prepare.soil.triangle.R
\name{soil.final.cleanup}
\alias{soil.final.cleanup}
\title{Prepare soil data to make figures}
\usage{
soil.final.cleanup(paired_data, first = FALSE)

soil.final.cleanup(paired_data, first = FALSE)
}
\arguments{
\item{paired_data}{}

\item{data}{}
}
\value{
Data frame with soil data

Data frame with soil data
}
\description{
For a dataframe output from function pair.data(), produces a table that shows 2018 data and amount of change.

This table is designed to be the source data for our functions that make figures

For a dataframe output from function pair.data(), produces a table that shows 2018 data and amount of change.

This table is designed to be the source data for our functions that make figures

prepare.soil.triangle() takes the result of soil.final.cleanup and adds necessary columns to make a soil triangle
}
\examples{
soil.final.cleanup(soildata)

soil.final.cleanup(soildata)
prepare.soil.triangle(soil)

}
