% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_species.cover.table.R
\name{species.cover.table}
\alias{species.cover.table}
\title{Create a table of cover for each species at each point}
\usage{
species.cover.table(lpi, transect, type = "absolute",
  surveyyear = max(levels(as.factor(lpi$year))))
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{transect}{Ranch for which to summarize cover}

\item{type}{Denotes type of cover to calculate"absolute" or "relative"}

\item{surveyyear}{The year for which to generate the summary}
}
\value{
A dataframe summary of cover of species
}
\description{
For veg data, summarizes cover of each species at each point.
}
\examples{
data = species.cover.table(lpi, type = "relative")

}
