% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_species.list.R
\name{species.list}
\alias{species.list}
\title{Create species list for veg data}
\usage{
species.list(lpi, releve, transect,
  surveyyear = levels(as.factor(lpi$year)))
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{releve}{A dataframe object of releve data from a veg survey}

\item{transect}{Ranch for which to generate species list}

\item{surveyyear}{Year for which to generate list}
}
\value{
A species list for a given dataset
}
\description{
For veg data, summarizes cover of shrubs, trees, litter, thatch, bare ground. Tree and shrub covers come from releve estimates
}
\examples{
data = species.list(lpi, checklist)

}
