% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_survey.sums.R, R/soil_survey_summary.R
\name{survey.sums}
\alias{survey.sums}
\title{Summarize data for soil survey visits}
\usage{
survey.sums(data)

survey.sums(data)
}
\arguments{
\item{data}{}

\item{data}{}
}
\value{
Data frame with One row per survey event, includes survey point names, date info, soil texture data, carbon % data, mean BD, and mean Infiltration time

Data frame with One row per survey event, includes survey point names, date info, soil texture data, carbon % data, mean BD, and mean Infiltration time
}
\description{
For a dataframe output from function add_soilcolumns(), produces a table that shows mean BD and Infiltration measurements

For a dataframe output from function add_soilcolumns(), produces a table that shows mean BD and Infiltration measurements
}
\examples{
survey_sums(soildata)

survey_sums(soildata)

}
