% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_texture.triangle.plot.R
\name{texture.triangle.plot}
\alias{texture.triangle.plot}
\title{Make a soil texture triangle}
\usage{
texture.triangle.plot(data, transect, labels = TRUE)
}
\arguments{
\item{labels}{}
}
\value{
Soil classification triangle with chosen ranch points emboldened

soil texture triangle
}
\description{
From output of prepare.soil.triangle() creates a soil texture plot and compares a ranch to other ranches in the dataset
}
\examples{
texture.triangle.plot(soil, "ranchname")


}
