% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_toSeconds.r, R/toSeconds.r
\name{toSeconds}
\alias{toSeconds}
\title{Convert to seconds}
\usage{
toSeconds(x)

toSeconds(x)
}
\arguments{
\item{x}{}

\item{x}{}
}
\value{
Time in seconds

Time in seconds
}
\description{
For infiltration times listed as character strings with the format HH:MM:SS return just seconds

For infiltration times listed as character strings with the format HH:MM:SS return just seconds
}
