% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_cleanup.R
\name{add.soilcolumns}
\alias{add.soilcolumns}
\title{Add date, year, and survey ID columns}
\usage{
add.soilcolumns(data, timeformat = "mm/dd/yyyy", inf_diameter = 15.2,
  bd_diameter = 5.2, volume = 450, bd_height = 7.5)
}
\arguments{
\item{bd_height}{}
}
\value{
data frame with added columns
}
\description{
For soil data, adds columns with date formatted as a date object in R, year as a numeric object, and Survey as a character object that inludes the point name and the survey year

Also adds columns that display BD and Infiltration diameter, height, and volume, and one that converts infiltration times to minutes shown as a base-10 decimal numeric
}
\examples{
data = add_datetime(data, timeformat = "yyyy-mm-dd")
x = include_datetime(soildata, timeformat = "mm/dd/yyyy")

}
