% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_Tidyname.R, R/all_Tidynames.R
\name{all_Tidynames}
\alias{all_Tidynames}
\title{Tidy column names in any data frame.}
\usage{
all_Tidynames(df)

all_Tidynames(df)
}
\arguments{
\item{df}{A dataframe object}

\item{df}{A dataframe object}
}
\value{
A data frame with tidy column names

A data frame with tidy column names
}
\description{
Column names are extremely important for programs to work smoothly. This function can be used during the first stage to standardize column names. This uses the make.names function and returns the new data frame with tidy names.

Column names are extremely important for programs to work smoothly. This function can be used during the first stage to standardize column names. This uses the make.names function and returns the new data frame with tidy names.
}
\examples{
data = all_Tidynames(data)

data = all_Tidynames(data)

}
