% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_carbon.plot.R
\name{carbon.plot}
\alias{carbon.plot}
\title{Carbon Plot}
\usage{
carbon.plot(data, transect, labels = TRUE)
}
\arguments{
\item{labels}{}
}
\value{
plot of soil carbon data
}
\description{
Plots soil carbon as scatterplot. Points from ranch identified are black, and all others are grayed out. Points are labeled automatically unless specified with labels = FALSE
}
\examples{
carbon.plot(soil, "ranchname")

}
