% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_compaction.plot.r
\name{compaction.plot}
\alias{compaction.plot}
\title{Make a soil compaction plot}
\usage{
compaction.plot(data, transect, year, labels = TRUE,
  pointcolors = c("black", "gray"), legend = FALSE,
  legendnames = c(transect, "Others"),
  xlab = "Bulk density, distance from target",
  ylab = "Infiltration, distance from target", box.padding = 0.5,
  linetype = "dashed", Inftarget = 3.81, xlims = c(NA, NA),
  ylims = c(NA, NA))
}
\arguments{
\item{data}{dataframe from output of soil.final.cleanup()}

\item{transect}{Character string of selected ranch code}

\item{labels}{= TRUE determines whether points from the chosen ranch are labeled on the plot}

\item{pointcolors}{= c("black", "gray") a vector that specifies colors of points on the plot}

\item{legend}{specifies whether to display a legend}

\item{legendnames}{specifies how points are named on the legend}

\item{xlab, ylab}{character strings for axis lables}

\item{box.padding}{numeric adjusts spacing of labels on the plot}

\item{linetype}{specifies type of line shown for BD and Infiltration targets}

\item{Inftarget}{target infiltration time}

\item{xlims, ylims}{numeric vectors of length 2 that specify the respective lower and upper limits of the axes. NA values cause default axis limits}
}
\value{
Shows soil compaction plot
}
\description{
From output of  creates a soil compaction plot displaying BD and Infiltration data, and compares to other ranches in the region
}
\examples{
compaction.plot(soil, "ranchname")


}
