% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_cover.summary.R
\name{cover.summary}
\alias{cover.summary}
\title{Create cover summary table for veg data}
\usage{
cover.summary(checklist, lpi, choose.variable = c("SpeciesRichness",
  "Litter", "Thatch", "BareGround", "Trees", "Shrubs"))
}
\arguments{
\item{checklist}{A dataframe object of checklist data from a veg survey}

\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{choose.variable}{Character vector that identifies which variables to summarize. Defaults to Species Richness, Litter, Thatch, Bare Ground, Trees, and Shrubs}
}
\value{
A summary of cover
}
\description{
For veg data, summarizes cover of shrubs, trees, litter, thatch, bare ground. Tree and shrub covers come from releve estimates
}
\examples{
data = cover.summary(checklist, lpi)

}
