% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_data.sum.R
\name{data.sum}
\alias{data.sum}
\title{Make a table to summarize data}
\usage{
data.sum(data, transect)
}
\arguments{
\item{transect}{}
}
\value{
Shows soil data in a nice-looking table

data summary table
}
\description{
From output of prepare.soil.triangle() creates a soil texture plot and compares a ranch to other ranches in the dataset
}
\examples{
data.sum(soil, "ranchname")


}
