% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_functional.cover.change.plot.R
\name{functional.cover.change.plot}
\alias{functional.cover.change.plot}
\title{Functional cover change summary plot}
\usage{
functional.cover.change.plot(lpi, transect, invasives = FALSE,
  surveyyear = levels(as.factor(lpi$year)), background = TRUE,
  type = "absolute", xlab = "Functional Group",
  ylab = paste("Cover Change", " ", min(levels(as.factor(surveyyear))),
  "-", max(levels(as.factor(surveyyear))), sep = ""),
  legendtitle = "Ranch", legendnames = c(paste(transect, collapse =
  " "), "Others"), boxcolors = c("black", "gray"))
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{transect}{Ranch for which to make plot}

\item{invasives}{Boolean whether to include invasives as a functional group}

\item{surveyyear}{The years for which to make the plot}

\item{background}{= TRUE whether to display "background" data on plot not from selected ranch}

\item{type}{"absolute" or "relative" cover}

\item{xlab, ylab}{Axis Labels}

\item{legendtitle}{Character string of legend title}

\item{legendnames}{specifies how points are named on the legend}

\item{boxcolors}{vector of colors to display data on the boxplot}
}
\value{
A ggplot of functional cover change with error bars
}
\description{
For veg data, summarizes change in mean cover of functional groups for a property

This will use data from the FIRST and LAST year specified for which there is lpi data on the selected ranch

Regional comparison data will use these same years to calculate average change

EX:: lpi data covers 2015-2019. Ranch XXXX has data from 2016 and 2019. Regional averages will therefore only include other properties that were sampled in 2016 and 2019. Ranches surveyed in 2015 and 2018 are not included in the regional calculation.
}
\examples{
data = functional.cover.plot(lpi, type = "absolute")

}
