% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_functional.cover.plot.R
\name{functional.cover.plot}
\alias{functional.cover.plot}
\title{Functional cover summary plot}
\usage{
functional.cover.plot(lpi, type = "absolute")
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey\}
}
\value{
A ggplot of functional cover with error bars
}
\description{
For veg data, summarizes and plots mean cover of functional groups for a property
}
\examples{
data = functional.cover.plot(lpi, type = "absolute")

}
