% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_functional.cover.plot.R
\name{functional.cover.plot}
\alias{functional.cover.plot}
\title{Functional cover summary plot}
\usage{
functional.cover.plot(lpi, type = "absolute", transect,
  surveyyear = max(levels(as.factor(lpi$year))),
  xlab = "Functional Group", ylab = "Percent Cover")
}
\arguments{
\item{lpi}{A dataframe object of lpi data from a veg survey}

\item{type}{"absolute" or "relative" cover}

\item{transect}{Ranch for which to make plot}

\item{surveyyear}{The year for which to calculate cover}

\item{xlab, ylab}{Axis labels}
}
\value{
A ggplot of functional cover with error bars
}
\description{
For veg data, summarizes and plots mean cover of functional groups for a property
}
\examples{
data = functional.cover.plot(lpi, type = "absolute")

}
