% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_percent.change.plot.R
\name{percent.change.plot}
\alias{percent.change.plot}
\title{Make a percent change boxplot of soil carbon and bulk Density}
\usage{
percent.change.plot(data, transect,
  surveyyear = levels(as.factor(data$YEAR)),
  legendnames = c(paste(transect, collapse = " "), "Others"),
  legendtitle = "Ranch", boxcolors = c("black", "gray"),
  xlabels = c("Carbon 0-10 cm", "Carbon 10-40 cm", "Bulk Density"),
  ylab = "\% Change", choosevariables = c("Carbon010change",
  "Carbon1040change", "Bulk.density.change"))
}
\arguments{
\item{data}{from output of soil.final.cleanup()}

\item{transect}{Character string of selected ranch code}

\item{surveyyear}{Years for which to make summary}

\item{legendnames}{specifies how points are named on the legend}

\item{legendtitle}{Character string of legend title}

\item{boxcolors}{vector of colors to display data on the boxplot}

\item{xlabels}{vector of strings to label boxplots}

\item{ylab}{character string for y axis label}

\item{choosevariables}{a character vector that specifies which columns in the data to plot}
}
\value{
Shows soil compaction plot
}
\description{
Creates a plot that shows average change in SOC at 2 depths and BD compared to regional average
}
\examples{
compaction.plot(soil, "ranchname")


}
