% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_remove_first.r
\name{remove_first}
\alias{remove_first}
\title{Remove first row of data.frame if empty}
\usage{
remove_first(df)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
A ggplot of functional cover with error bars
}
\description{
A reoccuring problem when importing data frames into rstudio is a blank first row that is added to the data frame. This function looks at the.
}
\examples{
data = functional.cover.plot(lpi, type = "absolute")

}
