% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_texture.triangle.plot.R
\name{texture.triangle.plot}
\alias{texture.triangle.plot}
\title{Make a soil texture triangle}
\usage{
texture.triangle.plot(data, transect, year, labels = TRUE, pch = 16,
  main = paste("Soil Texture - ", transect), cex.axis = 0.5,
  cex.lab = 0.75, cex.main = 0.75, text.tol = 0.2,
  tri.sum.tst = FALSE, cex = 0.75, frame.bg.col = "white",
  grid.show = FALSE)
}
\arguments{
\item{data}{dataframe, output of prepare.soil.triangle()}

\item{transect}{Character string of selected ranch code}

\item{labels}{= TRUE determines whether points from the chosen ranch are labeled on the plot}

\item{...}{arguments supplied to TT.plot}
}
\value{
Soil classification triangle with chosen ranch points emboldened

soil texture triangle
}
\description{
From output of prepare.soil.triangle() creates a soil texture plot and compares a ranch to other ranches in the dataset
}
\examples{
texture.triangle.plot(soil, "ranchname")


}
